# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Harden LLM API
  version: 1.0.0
  description: >-
    Provider-neutral self-hosted LLM runtime API. All application responses use
    a stable state/result/error envelope, except health probes and successful
    artifact redirects. Clients must not automatically retry an ambiguous run.
servers:
  - url: /
    description: Same-origin production gateway.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/state:
    get:
      operationId: getState
      tags: [State]
      summary: Read workspace state
      description: Returns the authenticated owner's persisted client-owned fields.
      responses:
        "200":
          $ref: "#/components/responses/StateSuccess"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: saveState
      tags: [State]
      summary: Save workspace state
      description: Strictly validates and replaces the authenticated owner's state document.
      requestBody:
        $ref: "#/components/requestBodies/ClientState"
      responses:
        "200":
          $ref: "#/components/responses/StateSuccess"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/profiles:
    get:
      operationId: listProfiles
      tags: [Profiles]
      summary: List redacted profiles
      description: Returns safe profile configuration and write-only credential status.
      responses:
        "200":
          $ref: "#/components/responses/ProfilesSuccess"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/history:
    get:
      operationId: listHistory
      tags: [Runs]
      summary: List run history
      description: Returns a stable owner-scoped page ordered by start time and run ID.
      parameters:
        - $ref: "#/components/parameters/HistoryCursor"
        - $ref: "#/components/parameters/HistoryLimit"
      responses:
        "200":
          $ref: "#/components/responses/HistorySuccess"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: clearHistory
      tags: [Runs]
      summary: Clear run history
      description: Deletes all owned runs, domain traces, observations, artifact metadata, and artifact object bodies.
      responses:
        "200":
          $ref: "#/components/responses/ClearHistorySuccess"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/history/{historyID}:
    parameters:
      - $ref: "#/components/parameters/HistoryID"
    delete:
      operationId: deleteHistory
      tags: [Runs]
      summary: Delete one history item
      description: Deletes an owned run, its domain trace, observations, artifact metadata, and artifact object bodies.
      responses:
        "200":
          $ref: "#/components/responses/DeleteHistorySuccess"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/stats:
    get:
      operationId: getStats
      tags: [Runs]
      summary: Read aggregate run statistics
      description: Computes authoritative owner-scoped totals directly from persisted run records.
      responses:
        "200":
          $ref: "#/components/responses/StatsSuccess"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/profiles/bundle:
    get:
      operationId: exportProfileBundle
      tags: [Profiles]
      summary: Export the profile bundle
      description: Returns profiles and opaque authenticated encrypted credentials.
      responses:
        "200":
          $ref: "#/components/responses/BundleSuccess"
        default:
          $ref: "#/components/responses/Error"
    put:
      operationId: importProfileBundle
      tags: [Profiles]
      summary: Replace the profile bundle
      description: Authenticates, probes, and atomically replaces the complete owner bundle.
      requestBody:
        $ref: "#/components/requestBodies/ProfileBundle"
      responses:
        "200":
          $ref: "#/components/responses/ProfilesSuccess"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/profiles/{profileID}:
    parameters:
      - $ref: "#/components/parameters/ProfileID"
    put:
      operationId: saveProfile
      tags: [Profiles]
      summary: Save a profile
      description: Validates and probes a profile before atomically persisting its credential.
      requestBody:
        $ref: "#/components/requestBodies/ProfileSave"
      responses:
        "200":
          $ref: "#/components/responses/ProfileSuccess"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteProfile
      tags: [Profiles]
      summary: Delete a profile
      description: Deletes an owned profile unless the remaining backup graph references it.
      responses:
        "200":
          $ref: "#/components/responses/DeleteProfileSuccess"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/profiles/{profileID}/models:refresh:
    parameters:
      - $ref: "#/components/parameters/ProfileID"
    post:
      operationId: refreshProfileModels
      tags: [Profiles]
      summary: Refresh profile models
      description: Discovers and normalizes models without replacing prior state on failure.
      responses:
        "200":
          $ref: "#/components/responses/ProfileSuccess"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/traces/{traceID}:
    parameters:
      - $ref: "#/components/parameters/TraceID"
    get:
      operationId: getTrace
      tags: [Traces]
      summary: Read a domain trace
      description: Returns one redacted trace with ordered observations, safe artifact metadata, and explicitly available persisted request/response resources.
      responses:
        "200":
          $ref: "#/components/responses/TraceSuccess"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/traces/{traceID}/artifacts/{artifactID}:
    parameters:
      - $ref: "#/components/parameters/TraceID"
      - $ref: "#/components/parameters/ArtifactID"
    get:
      operationId: getArtifact
      tags: [Traces]
      summary: Authorize an artifact download
      description: Authorizes owner metadata before issuing a short-lived presigned redirect.
      responses:
        "303":
          description: A short-lived URL for the authorized artifact.
          headers:
            Location:
              required: true
              description: Presigned artifact URL. Treat the complete value as sensitive.
              schema:
                type: string
                format: uri
              example: https://artifacts.example.test/harden-llm/example.json?X-Amz-Expires=60
        default:
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: A random opaque session token returned once by login, or the configured static CLI token.
  parameters:
    HistoryCursor:
      name: cursor
      in: query
      required: false
      description: Opaque stable pagination cursor returned by the prior page.
      schema:
        type: string
        minLength: 1
        maxLength: 512
    HistoryLimit:
      name: limit
      in: query
      required: false
      description: Maximum number of history items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    HistoryID:
      name: historyID
      in: path
      required: true
      description: Stable run identifier.
      schema:
        $ref: "#/components/schemas/Identifier"
    ProfileID:
      name: profileID
      in: path
      required: true
      description: UTF-8 profile identifier; the encoded path segment must not contain a slash.
      schema:
        $ref: "#/components/schemas/ProfileID"
    TraceID:
      name: traceID
      in: path
      required: true
      description: Stable trace identifier.
      schema:
        $ref: "#/components/schemas/Identifier"
    ArtifactID:
      name: artifactID
      in: path
      required: true
      description: Stable artifact identifier.
      schema:
        $ref: "#/components/schemas/Identifier"
  requestBodies:
...
    ClientState:
      required: true
      description: Complete client-owned state. Maximum encoded body size is 65536 bytes.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ClientState"
          examples:
            workspace:
              value:
                schemaVersion: 1
                selectedProfileId: Primary
                modelId: gpt-example
                userPrompt: Summarize the incident.
                callType: text
                structuredRepair: false
                cacheMode: "off"
    ProfileSave:
      required: true
      description: Profile and write-only credential. Maximum encoded body size is 262144 bytes.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProfileSaveRequest"
          examples:
            create:
              value:
                profile:
                  schemaVersion: 1
                  llmProfile: Primary
                  provider: openai
                  apiInferenceType: responses
                  endpointCredentialScope: user
                  baseUrl: https://api.openai.com/v1
                  modelId: gpt-example
                  pricing: null
                  supportsTemperature: true
                  supportsContractedStructuredOutput: true
                  tokensParam: max_output_tokens
                  responsesTokensParam: max_output_tokens
                  defaultOptions: {}
                  backupProfiles: []
                credentialId: primary-openai
                credential:
                  apiKey: example-write-only-key
    ProfileBundle:
      required: true
      description: Complete encrypted bundle. Maximum encoded body size is 2097152 bytes.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProfileBundle"
          examples:
            replace:
              value:
                schemaVersion: 1
                bundleId: bundle-example
                createdAt: "2026-07-13T12:00:00Z"
                profiles:
                  Primary:
                    schemaVersion: 1
                    llmProfile: Primary
                    provider: openai
                    apiInferenceType: responses
                    endpointCredentialScope: user
                    baseUrl: https://api.openai.com/v1
                    modelId: gpt-example
                    pricing: null
                    supportsTemperature: true
                    supportsContractedStructuredOutput: true
                    tokensParam: max_output_tokens
                    responsesTokensParam: max_output_tokens
                    defaultOptions: {}
                    backupProfiles: []
                credentialIds:
                  Primary: primary-openai
                credentials:
                  - schemaVersion: 1
                    binding:
                      ownerId: owner-example
                      credentialId: primary-openai
                      origin: https://api.openai.com
                    scope: user
                    apiInferenceTypes: [responses]
                    encrypted:
                      schemaVersion: 1
                      algorithm: AES-256-GCM
                      keyId: key-2026
                      nonce: AAAAAAAAAAAAAAAA
                      ciphertext: AAAAAAAAAAAAAAAAAAAAAA
                    createdAt: "2026-07-13T12:00:00Z"
                    updatedAt: "2026-07-13T12:00:00Z"
...
  responses:
...
    Error:
      description: Stable error envelope. Status is selected by the operation failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidRequest:
              value:
                state: {}
                result: null
                error:
                  code: invalid_request
                  message: The request is invalid.
...
    StateSuccess:
      description: The canonical current client state is carried in the state field.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StateEnvelope"
          examples:
            workspace:
              value:
                state:
                  schemaVersion: 1
                  selectedProfileId: Primary
                  modelId: gpt-example
                  userPrompt: Summarize the incident.
                  callType: text
                  structuredRepair: false
                  cacheMode: "off"
                result: null
                error: null
    ProfilesSuccess:
      description: Redacted profile states without encrypted or plaintext credential material.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProfilesEnvelope"
          examples:
            list:
              value:
                state: {}
                result:
                  profiles:
                    - profile:
                        schemaVersion: 1
                        llmProfile: Primary
                        provider: openai
                        apiInferenceType: responses
                        endpointCredentialScope: user
                        baseUrl: https://api.openai.com/v1
                        modelId: gpt-example
                        pricing: null
                        supportsTemperature: true
                        supportsContractedStructuredOutput: true
                        tokensParam: max_output_tokens
                        responsesTokensParam: max_output_tokens
                        defaultOptions: {}
                        backupProfiles: []
                      credential:
                        schemaVersion: 1
                        credentialId: primary-openai
                        scope: user
                        origin: https://api.openai.com
                        apiInferenceTypes: [responses]
                        configured: true
                        createdAt: "2026-07-13T12:00:00Z"
                        updatedAt: "2026-07-13T12:00:00Z"
                error: null
    ProfileSuccess:
      description: One redacted profile and its write-only credential status.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProfileEnvelope"
          examples:
            saved:
              value:
                state: {}
                result:
                  profile:
                    schemaVersion: 1
                    llmProfile: Primary
                    provider: openai
                    apiInferenceType: responses
                    endpointCredentialScope: user
                    baseUrl: https://api.openai.com/v1
                    modelId: gpt-example
                    pricing: null
                    supportsTemperature: true
                    supportsContractedStructuredOutput: true
                    tokensParam: max_output_tokens
                    responsesTokensParam: max_output_tokens
                    defaultOptions: {}
                    backupProfiles: []
                  credential:
                    schemaVersion: 1
                    credentialId: primary-openai
                    scope: user
                    origin: https://api.openai.com
                    apiInferenceTypes: [responses]
                    configured: true
                    createdAt: "2026-07-13T12:00:00Z"
                    updatedAt: "2026-07-13T12:00:00Z"
                error: null
    BundleSuccess:
      description: Complete profile catalog with authenticated encrypted credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BundleEnvelope"
          examples:
            export:
              value:
                state: {}
                result:
                  schemaVersion: 1
                  bundleId: bundle-example
                  createdAt: "2026-07-13T12:00:00Z"
                  profiles:
                    Primary:
                      schemaVersion: 1
                      llmProfile: Primary
                      provider: openai
                      apiInferenceType: responses
                      endpointCredentialScope: user
                      baseUrl: https://api.openai.com/v1
                      modelId: gpt-example
                      pricing: null
                      supportsTemperature: true
                      supportsContractedStructuredOutput: true
                      tokensParam: max_output_tokens
                      responsesTokensParam: max_output_tokens
                      defaultOptions: {}
                      backupProfiles: []
                  credentialIds: {Primary: primary-openai}
                  credentials:
                    - schemaVersion: 1
                      binding:
                        ownerId: owner-example
                        credentialId: primary-openai
                        origin: https://api.openai.com
                      scope: user
                      apiInferenceTypes: [responses]
                      encrypted:
                        schemaVersion: 1
                        algorithm: AES-256-GCM
                        keyId: key-2026
                        nonce: AAAAAAAAAAAAAAAA
                        ciphertext: AAAAAAAAAAAAAAAAAAAAAA
                      createdAt: "2026-07-13T12:00:00Z"
                      updatedAt: "2026-07-13T12:00:00Z"
                error: null
    HistorySuccess:
      description: Stable cursor page of owned run history.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HistoryEnvelope"
          examples:
            page:
              value:
                state: {}
                result:
                  items:
                    - runId: run-example
                      profileId: Primary
                      traceId: trace-example
                      status: succeeded
                      request: {profileId: Primary, userPrompt: Reply with OK., callType: text}
                      result: {output: OK}
                      startedAt: "2026-07-13T12:00:00Z"
                      completedAt: "2026-07-13T12:00:01Z"
                  nextCursor: eyJzdGFydGVkQXQiOiIyMDI2LTA3LTEzVDEyOjAwOjAwWiIsImlkIjoicnVuLWV4YW1wbGUifQ
                error: null
    StatsSuccess:
      description: Authoritative aggregates for all owned run records.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StatsEnvelope"
          examples:
            totals:
              value:
                state: {}
                result:
                  totalCount: 3
                  successCount: 2
                  failureCount: 1
                  timeoutCount: 0
                  totalPromptTokens: 24
                  cacheReadTokens: 8
                  cacheCreationTokens: 2
                  totalOutputTokens: 12
                  reasoningTokens: 4
                  totalTokens: 42
                  totalCost: 0.00042
                  cachedCost: 0.0001
                  cachedCount: 1
                  knownCostCount: 2
                  unknownCostCount: 1
                  totalCallDurationMs: 2580
                  maxCallDurationMs: 1200
                  overBudgetCount: 1
                  maxOverBudgetMs: 50
                error: null
    DeleteHistorySuccess:
      description: Confirmation that one run was deleted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DeleteHistoryEnvelope"
          examples:
            deleted:
              value:
                state: {}
                result: {deleted: true, runId: run-example}
                error: null
    ClearHistorySuccess:
      description: Count of deleted run records.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ClearHistoryEnvelope"
          examples:
            cleared:
              value:
                state: {}
                result: {deletedCount: 3}
                error: null
    DeleteProfileSuccess:
      description: Confirmation that one profile was deleted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DeleteProfileEnvelope"
          examples:
            deleted:
              value:
                state: {}
                result: {deleted: true, profileId: Primary}
                error: null
...
    TraceSuccess:
      description: Redacted domain trace, ordered observations, and safe artifact metadata.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TraceEnvelope"
          examples:
            trace:
              value:
                state: {}
                result:
                  traceId: trace-example
                  record: {schemaVersion: 1, status: succeeded, profileId: Primary}
                  observations:
                    - sequence: 0
                      type: provider.attempt
                      data: {number: 1, profileId: Primary}
                      createdAt: "2026-07-13T12:00:01Z"
                  artifacts:
                    - artifactId: call-example-trace
                      kind: trace
                      sha256: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
                      sizeBytes: 512
                      contentType: application/json
                      createdAt: "2026-07-13T12:00:01Z"
                  resources:
                    request:
                      available: true
                      payload: {profileId: Primary, userPrompt: "say ok", callType: text}
                    response:
                      available: true
                      payload: {output: "ok"}
                error: null
  schemas:
    Identifier:
      type: string
      minLength: 1
      maxLength: 128
      pattern: "^[A-Za-z0-9_-][A-Za-z0-9._-]{0,127}$"
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [state, result, error]
      properties:
        state: {$ref: "#/components/schemas/EmptyState"}
        result: {type: "null"}
        error: {$ref: "#/components/schemas/Error"}
...
    ClientState:
      type: object
      additionalProperties: false
      required: [schemaVersion, callType, structuredRepair, cacheMode]
      properties:
        schemaVersion: {type: integer, const: 1}
        selectedProfileId: {$ref: "#/components/schemas/ProfileID"}
        modelId: {type: string, maxLength: 512}
        systemPrompt: {type: string, maxLength: 32768}
        userPrompt: {type: string, maxLength: 65536}
        schemaShorthand: {type: string, maxLength: 65536}
        callType: {type: string, enum: [text, structured]}
        schema:
          oneOf:
            - {type: object}
            - {type: "null"}
        reasoningEffort: {type: string, enum: [lowest, middle, highest]}
        reasoningByProfile:
          type: object
          additionalProperties: {type: string, enum: [lowest, middle, highest]}
        structuredRepair: {type: boolean}
        cacheMode: {type: string, enum: ["off", cache, refresh]}
        providerOptions: {type: object}
        maxAttempts: {type: integer, minimum: 0, maximum: 10}
        initialBackoffMs: {type: integer, minimum: 0, maximum: 60000}
        maximumBackoffMs: {type: integer, minimum: 0, maximum: 600000}
        retryNetwork: {type: boolean}
        retryRateLimit: {type: boolean}
        retryServerError: {type: boolean}
        retryEmpty: {type: boolean}
        retryParse: {type: boolean}
        repairEscalation:
          oneOf:
            - type: object
              additionalProperties: false
              required: [attempt, modelId]
              properties:
                attempt: {type: integer, minimum: 2, maximum: 10}
                profileId: {$ref: "#/components/schemas/ProfileID"}
                modelId: {type: string, minLength: 1, maxLength: 512}
                reasoningEffort: {type: string, enum: [lowest, middle, highest]}
            - {type: "null"}
        ui: {type: object}
    StateEnvelope:
      type: object
      additionalProperties: false
      required: [state, result, error]
      properties:
        state: {$ref: "#/components/schemas/ClientState"}
        result: {type: "null"}
        error: {type: "null"}
...